backend/services/document_parser copy.py
"""
Enterprise-grade dynamic document parser - BULLETPROOF VERSION.

Uses ONLY your working packages: pdfplumber, tabula-py, easyocr, python-docx, pandas.
10+ formats with fallback chains, user isolation, deduplication.
"""

import os
import hashlib
import json
import pandas as pd
from pathlib import Path
from typing import List, Dict, Any, Optional
from dataclasses import dataclass,field
from datetime import datetime
import logging

# YOUR WORKING PACKAGES ONLY
import pdfplumber
import tabula
import easyocr
from docx import Document as DocxDocument

from core import get_logger, DocumentType, SUPPORTED_EXTENSIONS, DOCUMENTS_DIR
from core.constants import MAX_FILE_SIZE_MB
from config import get_settings

logger = get_logger(__name__)

# Global OCR reader (lazy init)
_ocr_reader = None


@dataclass
class ParsedDocument:
    document_id: str = ""
    filename: str = ""
    content: List[str] = field(default_factory=list)
    document_type: DocumentType = DocumentType.TXT
    metadata: Dict[str, Any] = field(default_factory=dict)
    pages: Optional[List[int]] = None
    user_id: Optional[str] = None
    file_hash: str = ""

def get_ocr_reader():
    """Lazy init EasyOCR."""
    global _ocr_reader
    if _ocr_reader is None:
        settings = get_settings()
        if settings.enable_ocr:
            try:
                _ocr_reader = easyocr.Reader(['en'], gpu=False)
                logger.info("EasyOCR initialized")
            except Exception as e:
                logger.warning(f"EasyOCR failed: {e}")
                _ocr_reader = None
    return _ocr_reader


class DocumentParser:
    """Bulletproof dynamic parser with fallback chains."""

    def __init__(self):
        self.settings = get_settings()
        self.supported_types = SUPPORTED_EXTENSIONS
        self.user_base_dirs = {
            "public": DOCUMENTS_DIR / "incoming" / "public",
            "private": DOCUMENTS_DIR / "incoming" / "private",
        }
        logger.info(f"Bulletproof Parser ready - {len(self.supported_types)} formats")

    def parse_file(
        self, 
        file_path: Path, 
        user_id: Optional[str] = None,
        is_public: bool = True
    ) -> Optional[ParsedDocument]:
        """Main parsing entry point."""
        if not file_path.exists():
            logger.error(f"File missing: {file_path}")
            return None

        file_size_mb = file_path.stat().st_size / (1024**2)
        if file_size_mb > MAX_FILE_SIZE_MB:
            logger.error(f"Too large ({file_size_mb:.1f}MB): {file_path.name}")
            return None

        ext = file_path.suffix.lower()
        if ext not in self.supported_types:
            logger.warning(f"Unsupported: {ext}")
            return None

        doc_type = self.supported_types[ext]
        doc_id = self._generate_id(file_path)
        file_hash = self._compute_hash(file_path)

        prefix = f"[{user_id or 'public'}]"
        logger.info(f"{prefix} Parsing {doc_type.value}: {file_path.name}")

        try:
            parsed = self._parse_dynamic(file_path, doc_type)
            if parsed:
                parsed.document_id = doc_id
                parsed.file_hash = file_hash
                parsed.user_id = user_id
                parsed.metadata = parsed.metadata or {}
                parsed.metadata.update({
                    "file_size_mb": round(file_size_mb, 2),
                    "processed_at": datetime.utcnow().isoformat()
                })
                logger.info(f"{prefix}  {len(parsed.content)} blocks")
                return parsed
        except Exception as e:
            logger.error(f"{prefix} Parse FAILED: {e}", exc_info=True)

        return None

    def _parse_dynamic(self, file_path: Path, doc_type: DocumentType) -> Optional[ParsedDocument]:
        """Dynamic parser dispatch with fallbacks."""
        parsers = {
            DocumentType.PDF: self._parse_pdf,
            DocumentType.DOCX: self._parse_docx,
            DocumentType.CSV: self._parse_csv,
            DocumentType.JSON: self._parse_json,
            DocumentType.TXT: self._parse_text,
            DocumentType.XLSX: self._parse_excel,
            DocumentType.PPTX: self._parse_pptx_fallback,
            DocumentType.HTML: self._parse_text,
            DocumentType.IMAGE: self._parse_image,
            DocumentType.MARKDOWN: self._parse_text,
        }
        return parsers.get(doc_type, self._parse_text)(file_path)

    # =========================================================================
    # PDF: pdfplumber → tabula → OCR
    # =========================================================================
    def _parse_pdf(self, file_path: Path) -> Optional[ParsedDocument]:
        """PDF: text + tables + OCR fallback."""
        content = []
        pages = []
        metadata = {"tables": 0}

        try:
            # 1. pdfplumber (text + layout)
            with pdfplumber.open(file_path) as pdf:
                for page_num, page in enumerate(pdf.pages, 1):
                    text = page.extract_text()
                    if text and text.strip():
                        content.append(text.strip())
                        pages.append(page_num)
                    
                    # Tables per page
                    tables = page.extract_tables()
                    for table in tables:
                        if table:
                            df = pd.DataFrame(table[1:], columns=table[0])
                            content.append(f"\n[TABLE p{page_num}]\n{df.to_markdown(index=False)}")
                            metadata["tables"] += 1

            if content:
                return ParsedDocument(
                    filename=file_path.name, content=content,
                    document_type=DocumentType.PDF, metadata=metadata, pages=pages
                )
        except Exception as e:
            logger.warning(f"pdfplumber failed: {e}")

        # 2. tabula fallback (tables only)
        try:
            tables = tabula.read_pdf(str(file_path), pages="all", multiple_tables=True)
            if tables:
                metadata["tables"] = len(tables)
                content = [f"[TABLE {i+1}]\n{df.to_markdown(index=False)}" for i, df in enumerate(tables)]
                return ParsedDocument(filename=file_path.name, content=content, document_type=DocumentType.PDF, metadata=metadata)
        except:
            pass

        # 3. OCR final fallback
        return self._parse_image(file_path)

    # =========================================================================
    # DOCX: python-docx (structure preserved)
    # =========================================================================
    def _parse_docx(self, file_path: Path) -> Optional[ParsedDocument]:
        """DOCX with headings + tables."""
        try:
            doc = DocxDocument(file_path)
            content = []
            metadata = {"headings": 0, "tables": 0, "paragraphs": 0}

            for para in doc.paragraphs:
                text = para.text.strip()
                if text:
                    if para.style.name.startswith('Heading'):
                        content.append(f"\n## {text}\n")
                        metadata["headings"] += 1
                    else:
                        content.append(text)
                    metadata["paragraphs"] += 1

            for table_idx, table in enumerate(doc.tables):
                table_md = self._docx_table_to_md(table)
                if table_md:
                    content.append(f"\n[TABLE {table_idx+1}]\n{table_md}\n")
                    metadata["tables"] += 1

            return ParsedDocument(
                filename=file_path.name, content=content,
                document_type=DocumentType.DOCX, metadata=metadata
            )
        except Exception:
            return self._parse_text(file_path)

    # =========================================================================
    # SIMPLE FORMATS
    # =========================================================================
    def _parse_csv(self, file_path: Path) -> Optional[ParsedDocument]:
        try:
            df = pd.read_csv(file_path)
            content = [df.to_markdown(index=False)]
            return ParsedDocument(
                filename=file_path.name, content=content,
                document_type=DocumentType.CSV,
                metadata={"rows": len(df), "columns": len(df.columns)}
            )
        except:
            return self._parse_text(file_path)

    def _parse_json(self, file_path: Path) -> Optional[ParsedDocument]:
        try:
            with open(file_path, 'r', encoding='utf-8') as f:
                data = json.load(f)
            content = [json.dumps(data, indent=2, ensure_ascii=False)]
            return ParsedDocument(
                filename=file_path.name, content=content,
                document_type=DocumentType.JSON,
                metadata={"keys": len(data) if isinstance(data, dict) else 0}
            )
        except:
            return self._parse_text(file_path)

    def _parse_excel(self, file_path: Path) -> Optional[ParsedDocument]:
        try:
            xl = pd.ExcelFile(file_path)
            content = []
            for sheet in xl.sheet_names:
                df = pd.read_excel(file_path, sheet_name=sheet)
                content.append(f"\n[SHEET: {sheet}]\n{df.to_markdown(index=False)}")
            return ParsedDocument(
                filename=file_path.name, content=content,
                document_type=DocumentType.XLSX,
                metadata={"sheets": len(xl.sheet_names)}
            )
        except:
            return self._parse_text(file_path)

    def _parse_text(self, file_path: Path) -> Optional[ParsedDocument]:
        try:
            with open(file_path, 'r', encoding='utf-8') as f:
                lines = [line.strip() for line in f if line.strip()]
            return ParsedDocument(
                filename=file_path.name, content=lines,
                document_type=DocumentType.TXT
            )
        except:
            return None

    def _parse_pptx_fallback(self, file_path: Path) -> Optional[ParsedDocument]:
        """PPTX text fallback."""
        return self._parse_text(file_path)

    def _parse_image(self, file_path: Path) -> Optional[ParsedDocument]:
        """EasyOCR images."""
        reader = get_ocr_reader()
        if not reader:
            return None

        try:
            results = reader.readtext(str(file_path))
            text = " ".join([result[1] for result in results])
            if text.strip():
                return ParsedDocument(
                    filename=file_path.name, content=[text.strip()],
                    document_type=DocumentType.IMAGE,
                    metadata={"ocr_results": len(results)}
                )
        except Exception as e:
            logger.error(f"OCR failed: {e}")
        return None

    # =========================================================================
    # HELPERS
    # =========================================================================
    def _docx_table_to_md(self, table) -> str:
        """DOCX table → markdown."""
        rows = []
        for row in table.rows:
            cells = [cell.text.strip() for cell in row.cells]
            rows.append("| " + " | ".join(cells) + " |")
        return "\n".join(rows) if rows else ""

    def _generate_id(self, file_path: Path) -> str:
        return hashlib.md5(
            f"{file_path.name}_{file_path.stat().st_size}_{file_path.stat().st_mtime}".encode()
        ).hexdigest()

    def _compute_hash(self, file_path: Path) -> str:
        h = hashlib.sha256()
        with open(file_path, "rb") as f:
            while chunk := f.read(8192):
                h.update(chunk)
        return h.hexdigest()
